최종본

+++ src/main/java/com/truthify/ad/dto/AdAnalyzeResponse.java
@@ -11,10 +11,8 @@ import lombok.AllArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AdAnalyzeResponse {
-	private Long adTextId; // 저장된 광고 텍스트 ID (PK)
-	private Integer riskScore; // AI가 판단한 위험 점수 (0-100)
-	private String summary; // 분석 결과 요약 설명
+	private Long adTextId;
+	private Double riskScore; 
+	private String summary; 
 	private String detailDescription;
-	// 추후 상세 분석 결과 필드 (e.g., exaggerationCount, highRiskSentences) 추가 예정
-
 }
